move client side signaling  logic to separate file
